pyproject.toml
[tool.poetry]
name = "tasklens-api"
version = "0.1.0"
description = "APIs for monitoring cronjobs computations"
authors = ["Supreeth Kunder <kundersupreeth@gmail.com>"]
readme = "README.md"

packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.112.0"}
python-dotenv = "^1.0.1"
pytest = "^8.3.2"
redis = "^5.0.8"
coloredlogs = "^15.0.1"
python-jose = "^3.3.0"
python-jwt = "^4.1.0"
jwt = "^1.3.1"
requests = "^2.32.3"
sse-starlette = "^2.1.3"
websockets = "11"
supabase = "^2.7.4"

[tool.poetry.scripts]
api = "src.cli:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
